requirements.txt
aiofiles==23.1.0
aiogram==3.0.0b7
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
black==23.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
flake8==6.0.0
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.2.0
pycodestyle==2.10.0
pydantic==1.10.5
pyflakes==3.0.1
python-decouple==3.8
python-dotenv==1.0.0
pytz==2022.7.1
typing_extensions==4.5.0
yarl==1.8.2
